S60: start sandbagging to keep latency down
Depending on the driver, there is little or no control about the output
buffer size. This adds up to several hundred milliseconds on my test phone
(Nokia E63).

This patch adds a delay before submitting samples for playing. If the
system's output counter is too far behind the submitted samples, it waits
until it isn't. This allows some measure of control over the invisible
buffers.

Of course, this introduces a magic value for the lag tolerance. Setting it
too low causes glitches. These tend to be evenly periodic - so maybe
rejigging things to submit (say) precisely 256 samples a throw would allow
pushing this further.